# docdoku-server/docdoku-server-web/src/main/webapp/swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "v2.5",
    "title" : "DocdokuPLM API"
  },
  "basePath" : "/api",
...
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/workspaces/{workspaceId}/document-templates" : {
      "get" : {
        "tags" : [ "document-templates", "workspaces" ],
        "summary" : "Get document templates",
        "description" : "",
        "operationId" : "getDocumentMasterTemplates",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/DocumentMasterTemplateDTO"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "document-templates", "workspaces" ],
        "summary" : "Create document template",
        "description" : "",
        "operationId" : "createDocumentMasterTemplate",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/DocumentMasterTemplateDTO"
            }
          }
        }
      }
    },
...
    "/workspaces/{workspaceId}/folders" : {
      "get" : {
        "tags" : [ "folders", "workspaces" ],
        "summary" : "Get root folders",
        "description" : "",
        "operationId" : "getRootFolders",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "configSpec",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/FolderDTO"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "folders", "workspaces" ],
        "summary" : "Create root folder",
        "description" : "",
        "operationId" : "createRootFolder",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/FolderDTO"
            }
          }
        }
      }
    },
...
    "/workspaces/{workspaceId}/parts" : {
      "get" : {
        "tags" : [ "parts", "workspaces" ],
        "summary" : "Get part revisions",
        "description" : "",
        "operationId" : "getPartRevisions",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "start",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "length",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      },
      "post" : {
        "tags" : [ "parts", "workspaces" ],
        "summary" : "Create new part",
        "description" : "",
        "operationId" : "createNewPart",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/PartRevisionDTO"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "ACLDTO" : {
      "type" : "object",
      "properties" : {
        "userEntries" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "string",
            "enum" : [ "FORBIDDEN", "READ_ONLY", "FULL_ACCESS" ]
          }
        },
        "groupEntries" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "string",
            "enum" : [ "FORBIDDEN", "READ_ONLY", "FULL_ACCESS" ]
          }
        }
      }
    },
...
    "ActivityDTO" : {
      "type" : "object",
      "properties" : {
        "step" : {
          "type" : "integer",
          "format" : "int32"
        },
        "relaunchStep" : {
          "type" : "integer",
          "format" : "int32"
        },
        "tasks" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/TaskDTO"
          }
        },
        "lifeCycleState" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string",
          "enum" : [ "SERIAL", "PARALLEL" ]
        },
        "tasksToComplete" : {
          "type" : "integer",
          "format" : "int32"
        },
        "complete" : {
          "type" : "boolean",
          "default" : false
        },
        "stopped" : {
          "type" : "boolean",
          "default" : false
        },
        "inProgress" : {
          "type" : "boolean",
          "default" : false
        },
        "toDo" : {
          "type" : "boolean",
          "default" : false
        }
      }
    },
...
    "CADInstanceDTO" : {
      "type" : "object",
      "properties" : {
        "rx" : {
          "type" : "number",
          "format" : "double"
        },
        "ry" : {
          "type" : "number",
          "format" : "double"
        },
        "rz" : {
          "type" : "number",
          "format" : "double"
        },
        "tx" : {
          "type" : "number",
          "format" : "double"
        },
        "ty" : {
          "type" : "number",
          "format" : "double"
        },
        "tz" : {
          "type" : "number",
          "format" : "double"
        },
        "matrix" : {
          "type" : "array",
          "items" : {
            "type" : "number",
            "format" : "double"
          }
        },
        "rotationType" : {
          "type" : "string",
          "enum" : [ "ANGLE", "MATRIX" ]
        }
      }
    },
    "ComponentDTO" : {
      "type" : "object",
      "properties" : {
        "author" : {
          "type" : "string"
        },
        "authorLogin" : {
          "type" : "string"
        },
        "number" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "version" : {
          "type" : "string"
        },
        "iteration" : {
          "type" : "integer",
          "format" : "int32"
        },
        "description" : {
          "type" : "string"
        },
        "standardPart" : {
          "type" : "boolean",
          "default" : false
        },
        "assembly" : {
          "type" : "boolean",
          "default" : false
        },
        "substitute" : {
          "type" : "boolean",
          "default" : false
        },
        "partUsageLinkId" : {
          "type" : "string"
        },
        "partUsageLinkReferenceDescription" : {
          "type" : "string"
        },
        "components" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ComponentDTO"
          }
        },
        "amount" : {
          "type" : "number",
          "format" : "double"
        },
        "unit" : {
          "type" : "string"
        },
        "attributes" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/InstanceAttributeDTO"
          }
        },
        "checkOutUser" : {
          "$ref" : "#/definitions/UserDTO"
        },
        "checkOutDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "released" : {
          "type" : "boolean",
          "default" : false
        },
        "obsolete" : {
          "type" : "boolean",
          "default" : false
        },
        "optional" : {
          "type" : "boolean",
          "default" : false
        },
        "lastIterationNumber" : {
          "type" : "integer",
          "format" : "int32"
        },
        "accessDeny" : {
          "type" : "boolean",
          "default" : false
        },
        "substituteIds" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "notifications" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ModificationNotificationDTO"
          }
        },
        "hasPathData" : {
          "type" : "boolean",
          "default" : false
        },
        "path" : {
          "type" : "string"
        },
        "virtual" : {
          "type" : "boolean",
          "default" : false
        }
      }
    },
...
    "DocumentIterationDTO" : {
      "type" : "object",
      "properties" : {
        "workspaceId" : {
          "type" : "string"
        },
        "id" : {
          "type" : "string"
        },
        "documentMasterId" : {
          "type" : "string"
        },
        "version" : {
          "type" : "string"
        },
        "iteration" : {
          "type" : "integer",
          "format" : "int32"
        },
        "creationDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "modificationDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "checkInDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "title" : {
          "type" : "string"
        },
        "author" : {
          "$ref" : "#/definitions/UserDTO"
        },
        "revisionNote" : {
          "type" : "string"
        },
        "attachedFiles" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "instanceAttributes" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/InstanceAttributeDTO"
          }
        },
        "linkedDocuments" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/DocumentRevisionDTO"
          }
        },
        "documentRevision" : {
          "$ref" : "#/definitions/DocumentRevisionDTO"
        }
      }
    },
...
    "DocumentMasterTemplateDTO" : {
      "type" : "object",
      "properties" : {
        "workspaceId" : {
          "type" : "string"
        },
        "id" : {
          "type" : "string"
        },
        "documentType" : {
          "type" : "string"
        },
        "author" : {
          "$ref" : "#/definitions/UserDTO"
        },
        "creationDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "modificationDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "idGenerated" : {
          "type" : "boolean",
          "default" : false
        },
        "mask" : {
          "type" : "string"
        },
        "workflowModelId" : {
          "type" : "string"
        },
        "attachedFiles" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "attributeTemplates" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/InstanceAttributeTemplateDTO"
          }
        },
        "attributesLocked" : {
          "type" : "boolean",
          "default" : false
        },
        "acl" : {
          "$ref" : "#/definitions/ACLDTO"
        }
      }
    },
...
    "DocumentRevisionDTO" : {
      "type" : "object",
      "properties" : {
        "workspaceId" : {
          "type" : "string"
        },
        "id" : {
          "type" : "string"
        },
        "documentMasterId" : {
          "type" : "string"
        },
        "version" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string"
        },
        "author" : {
          "$ref" : "#/definitions/UserDTO"
        },
        "creationDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "commentLink" : {
          "type" : "string"
        },
        "title" : {
          "type" : "string"
        },
        "checkOutUser" : {
          "$ref" : "#/definitions/UserDTO"
        },
        "checkOutDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "tags" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "description" : {
          "type" : "string"
        },
        "iterationSubscription" : {
          "type" : "boolean",
          "default" : false
        },
        "stateSubscription" : {
          "type" : "boolean",
          "default" : false
        },
        "documentIterations" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/DocumentIterationDTO"
          }
        },
        "workflow" : {
          "$ref" : "#/definitions/WorkflowDTO"
        },
        "path" : {
          "type" : "string"
        },
        "lifeCycleState" : {
          "type" : "string"
        },
        "publicShared" : {
          "type" : "boolean",
          "default" : false
        },
        "attributesLocked" : {
          "type" : "boolean",
          "default" : false
        },
        "acl" : {
          "$ref" : "#/definitions/ACLDTO"
        },
        "lastIteration" : {
          "$ref" : "#/definitions/DocumentIterationDTO"
        }
      }
    },
...
    "FolderDTO" : {
      "type" : "object",
      "properties" : {
        "path" : {
          "type" : "string"
        },
        "id" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "home" : {
          "type" : "boolean",
          "default" : false
        }
      }
    },
...
    "InstanceAttributeDTO" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "mandatory" : {
          "type" : "boolean",
          "default" : false
        },
        "locked" : {
          "type" : "boolean",
          "default" : false
        },
        "type" : {
          "type" : "string",
          "enum" : [ "TEXT", "NUMBER", "DATE", "BOOLEAN", "URL", "LOV" ]
        },
        "value" : {
          "type" : "string"
        },
        "lovName" : {
          "type" : "string"
        },
        "items" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/NameValuePairDTO"
          }
        }
      }
    },
...
    "InstanceAttributeTemplateDTO" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "mandatory" : {
          "type" : "boolean",
          "default" : false
        },
        "attributeType" : {
          "type" : "string",
          "enum" : [ "TEXT", "NUMBER", "DATE", "BOOLEAN", "URL", "LOV" ]
        },
        "lovName" : {
          "type" : "string"
        },
        "locked" : {
          "type" : "boolean",
          "default" : false
        }
      }
    },
...
    "ModificationNotificationDTO" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "impactedPartNumber" : {
          "type" : "string"
        },
        "impactedPartVersion" : {
          "type" : "string"
        },
        "modifiedPartNumber" : {
          "type" : "string"
        },
        "modifiedPartName" : {
          "type" : "string"
        },
        "modifiedPartVersion" : {
          "type" : "string"
        },
        "modifiedPartIteration" : {
          "type" : "integer",
          "format" : "int32"
        },
        "checkInDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "iterationNote" : {
          "type" : "string"
        },
        "author" : {
          "$ref" : "#/definitions/UserDTO"
        },
        "acknowledged" : {
          "type" : "boolean",
          "default" : false
        },
        "ackComment" : {
          "type" : "string"
        },
        "ackAuthor" : {
          "$ref" : "#/definitions/UserDTO"
        },
        "ackDate" : {
          "type" : "string",
          "format" : "date-time"
        }
      }
    },
    "NameValuePairDTO" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "value" : {
          "type" : "string"
        }
      }
    },
...
    "PartIterationDTO" : {
      "type" : "object",
      "properties" : {
        "workspaceId" : {
          "type" : "string"
        },
        "iteration" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nativeCADFile" : {
          "type" : "string"
        },
        "iterationNote" : {
          "type" : "string"
        },
        "author" : {
          "$ref" : "#/definitions/UserDTO"
        },
        "creationDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "modificationDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "checkInDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "instanceAttributes" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/InstanceAttributeDTO"
          }
        },
        "instanceAttributeTemplates" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/InstanceAttributeTemplateDTO"
          }
        },
        "components" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/PartUsageLinkDTO"
          }
        },
        "linkedDocuments" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/DocumentRevisionDTO"
          }
        },
        "number" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "version" : {
          "type" : "string"
        },
        "attachedFiles" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        }
      }
    },
...
    "PartRevisionDTO" : {
      "type" : "object",
      "properties" : {
        "partKey" : {
          "type" : "string"
        },
        "number" : {
          "type" : "string"
        },
        "version" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "author" : {
          "$ref" : "#/definitions/UserDTO"
        },
        "creationDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "modificationDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "checkInDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "description" : {
          "type" : "string"
        },
        "partIterations" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/PartIterationDTO"
          }
        },
        "checkOutUser" : {
          "$ref" : "#/definitions/UserDTO"
        },
        "checkOutDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "workflow" : {
          "$ref" : "#/definitions/WorkflowDTO"
        },
        "lifeCycleState" : {
          "type" : "string"
        },
        "standardPart" : {
          "type" : "boolean",
          "default" : false
        },
        "workspaceId" : {
          "type" : "string"
        },
        "publicShared" : {
          "type" : "boolean",
          "default" : false
        },
        "acl" : {
          "$ref" : "#/definitions/ACLDTO"
        },
        "attributesLocked" : {
          "type" : "boolean",
          "default" : false
        },
        "status" : {
          "type" : "string",
          "enum" : [ "WIP", "RELEASED", "OBSOLETE" ]
        },
        "lastIterationNumber" : {
          "type" : "integer",
          "format" : "int32"
        },
        "tags" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "notifications" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ModificationNotificationDTO"
          }
        },
        "obsoleteDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "obsoleteAuthor" : {
          "$ref" : "#/definitions/UserDTO"
        },
        "releaseDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "releaseAuthor" : {
          "$ref" : "#/definitions/UserDTO"
        }
      }
    },
...
    "PartSubstituteLinkDTO" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "fullId" : {
          "type" : "string"
        },
        "amount" : {
          "type" : "number",
          "format" : "double"
        },
        "unit" : {
          "type" : "string"
        },
        "referenceDescription" : {
          "type" : "string"
        },
        "comment" : {
          "type" : "string"
        },
        "substitute" : {
          "$ref" : "#/definitions/ComponentDTO"
        },
        "cadInstances" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/CADInstanceDTO"
          }
        }
      }
    },
...
    "PartUsageLinkDTO" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "fullId" : {
          "type" : "string"
        },
        "amount" : {
          "type" : "number",
          "format" : "double"
        },
        "comment" : {
          "type" : "string"
        },
        "component" : {
          "$ref" : "#/definitions/ComponentDTO"
        },
        "referenceDescription" : {
          "type" : "string"
        },
        "unit" : {
          "type" : "string"
        },
        "optional" : {
          "type" : "boolean",
          "default" : false
        },
        "cadInstances" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/CADInstanceDTO"
          }
        },
        "substitutes" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/PartSubstituteLinkDTO"
          }
        }
      }
    },
...
    "TaskDTO" : {
      "type" : "object",
      "properties" : {
        "closureComment" : {
          "type" : "string"
        },
        "title" : {
          "type" : "string"
        },
        "instructions" : {
          "type" : "string"
        },
        "targetIteration" : {
          "type" : "integer",
          "format" : "int32"
        },
        "closureDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "signature" : {
          "type" : "string"
        },
        "worker" : {
          "$ref" : "#/definitions/UserDTO"
        },
        "status" : {
          "type" : "string",
          "enum" : [ "NOT_STARTED", "IN_PROGRESS", "APPROVED", "REJECTED", "NOT_TO_BE_DONE" ]
        }
      }
    },
...
    "UserDTO" : {
      "type" : "object",
      "properties" : {
        "workspaceId" : {
          "type" : "string"
        },
        "login" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "language" : {
          "type" : "string"
        },
        "membership" : {
          "type" : "string",
          "enum" : [ "READ_ONLY", "FULL_ACCESS" ]
        }
      }
    },
...
    "WorkflowDTO" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "finalLifeCycleState" : {
          "type" : "string"
        },
        "activities" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ActivityDTO"
          }
        },
        "abortedDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "lifeCycleState" : {
          "type" : "string"
        },
        "currentActivity" : {
          "$ref" : "#/definitions/ActivityDTO"
        },
        "currentStep" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
...
  }
}